test: split isURL validator tests
